Add JavaScript files for modular functionallity
